Initial commit of a C++ symbolic execution library.
This source code repository features an extensible proof-of-concept for
a symbolic execution tool that uses operator overloading and reflection
to gather symbolic constraints along a concrete execution path. This
design supports a range of C++ language features by integrating with
existing compiler infrastructures. To use the library in a program under
test, we can use the compiler's pre-processing capabilities to redefine
the types (e.g. #define bool Bool). If the program under test has also
external open-source dependencies, the signatures of these library
functions could be modified to take symbolic expression types as
arguments (e.g. Bool). If the source code of such libraries is
unavailable or its modification is infeasible, precision is lost but it
should continue to compile due to implicit primitive type conversion
operators on symbolic expressions.

The unit and functional tests of this proof-of-concept can be invoked on
the command line by "make check".