Replace BinaryExpr with NaryExpr.
NaryExpr can reduce the depth of the DAG. The deeper this recursive data
structure is the more likely we are to exceed the call stack size during
the deallocation of smart pointers. Thanks to Michael Tautschnig who
helped bring attention to this problem.

Each NaryExpr has attributes associated with it that describe the
operator's mathematical properties (e.g. associativity). Future work is
required to incorporate type information (e.g. addition of integers is
associative whereas floating point addition is typically not).