Allow symbolic expressions to be reset.
Relax restrictions on Value<T>::set_symbolic(const std::string&). With
each call, it resets the symbolic expression to either a ValueExpr<T> or
an AnyExpr<T> object. More precisely, it is a ValueExpr<T> if directed
symbolic execution (i.e. DART-style symbolic execution) is supported;
otherwise, it instantiates a new AnyExpr<T> object. This behaviour is
useful for the symbolic execution of thread context switches.