Adopt "data" and "aggregate" terminology.
To avoid confusion with the terms symbolic and concolic "values", we call
the concrete data associated with a Value<T> object simply "data". If
nothing else, this avoids name clashes in the API, e.g.
Value<T>::data() vs Var<T>::value().

An "aggregate" value refers to the data that is used for constant propagation.