Delete original sequential symbolic execution code
Recall that the original source code is solely about sequential symbolic
execution. It allowed us to explore the design space of various algorithms
and data structures in the simpler scenario where there is only a single
thread of execution in the program under scrutiny.

In contrast, the new code targets the analysis of concurrent programs.
To achieve this, it has as fundamental building blocks data structures
for memory read and writes, i.e. events. In addition, there are now ways
to preserve the structure of programs without goto statements through a
series-parallel graph. Finally, the new library integrates more seamlessly
the C++11 compiler and Z3 theorem solver. These and other improvements
warrant the removal of the old code. While a few features -- such as
efficient constant propagation in atomic blocks -- still have to be
back-ported, these are not on the critical path to release usable code.