Treat MemoryAddr as a lattice element
The treatment of MemoryAddr as elements in a lattice means that these
are partially ordered. Denote this partial order by <=. Let a and a' be
MemoryAddr objects. Let e and e' be events such that e.addr() == a and
e'.addr() == a'. Informally, a <= a' iff e' affects less memory than e.

Changes:

* Add MemoryAddr::meet(const MemoryAddr&) and join(const MemoryAddr&)
* Deprecate static MemoryAddr::join(const MemoryAddr&, const MemoryAddr&)