Change Event encoder dispatcher functions
* Rename Event::encode(const Z3ValueEncoder&, Z3&) to Event::encode_eq(...)
* Add Event::constant(Z3&)

The new dispatcher function Event::constant(Z3&) is needed to allow the
Encoder create an appropriate Z3 constant from a given event as this is
often not possible by merely inspecting Event::type().

To see this, consider IndirectWriteEvent<Range, Domain, N> where Range
could be an integral type. While Event::type() captures the Range type
attributes of the memory being indirectly written, the event must be
encoded as an array constant. A similar example is ReadEvent<T[N]> that
is also encoded as an array constant.